@@ -15,7 +15,7 @@ apply plugin: "com.android.application"
 apply plugin: 'io.fabric'
 
 android {
-    compileSdkVersion 25
+    compileSdkVersion 28
     buildToolsVersion '27.0.1'
 
     signingConfigs {
