@@ -23,35 +23,51 @@ import android.app.DialogFragment;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
+import android.os.Environment;
+import android.os.Parcelable;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.MenuItem;
 
 import com.google.android.gms.cast.MediaInfo;
+import com.google.android.gms.cast.MediaLoadOptions;
 import com.google.android.gms.cast.MediaMetadata;
 import com.google.android.gms.cast.framework.CastButtonFactory;
 import com.google.android.gms.cast.framework.CastContext;
 import com.google.android.gms.cast.framework.CastSession;
 import com.google.android.gms.cast.framework.SessionManagerListener;
 import com.google.android.gms.cast.framework.media.RemoteMediaClient;
+import com.google.android.gms.common.images.WebImage;
 import com.squareup.otto.Subscribe;
 
 import org.amahi.anywhere.AmahiApplication;
 import org.amahi.anywhere.R;
 import org.amahi.anywhere.adapter.ServerFilesImagePagerAdapter;
 import org.amahi.anywhere.bus.BusProvider;
+import org.amahi.anywhere.bus.FileCopiedEvent;
 import org.amahi.anywhere.bus.FileDownloadedEvent;
+import org.amahi.anywhere.db.entities.OfflineFile;
+import org.amahi.anywhere.db.entities.RecentFile;
+import org.amahi.anywhere.db.repositories.OfflineFileRepository;
+import org.amahi.anywhere.db.repositories.RecentFileRepository;
+import org.amahi.anywhere.fragment.PrepareDialogFragment;
 import org.amahi.anywhere.fragment.ServerFileDownloadingFragment;
+import org.amahi.anywhere.model.FileOption;
 import org.amahi.anywhere.server.client.ServerClient;
 import org.amahi.anywhere.server.model.ServerFile;
 import org.amahi.anywhere.server.model.ServerShare;
+import org.amahi.anywhere.util.Downloader;
+import org.amahi.anywhere.util.FileManager;
 import org.amahi.anywhere.util.FullScreenHelper;
 import org.amahi.anywhere.util.Intents;
+import org.amahi.anywhere.util.Preferences;
 import org.amahi.anywhere.view.ClickableViewPager;
 
+import java.io.File;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -96,11 +112,30 @@ public class ServerFileImageActivity extends AppCompatActivity implements
 
         setUpHomeNavigation();
 
+        prepareFiles();
+
         setUpImage();
+
+        setUpRecentFiles(getFile());
 
         setUpCast();
 
         setUpFullScreen();
+    }
+
+    private void prepareFiles() {
+        int fileType = getIntent().getIntExtra(Intents.Extras.FILE_TYPE, FileManager.SERVER_FILE);
+        if (fileType == FileManager.RECENT_FILE) {
+            RecentFileRepository repository = new RecentFileRepository(this);
+            RecentFile recentFile = repository.getRecentFile(getIntent().getStringExtra(Intents.Extras.UNIQUE_KEY));
+            ServerFile serverFile = new ServerFile(recentFile.getName(), recentFile.getModificationTime(), recentFile.getMime());
+            serverFile.setSize(recentFile.getSize());
+            serverFile.setMime(recentFile.getMime());
+            getIntent().putExtra(Intents.Extras.SERVER_FILE, serverFile);
+            List<ServerFile> serverFiles = new ArrayList<>();
+            serverFiles.add(serverFile);
+            getIntent().putExtra(Intents.Extras.SERVER_FILES, new ArrayList<Parcelable>(serverFiles));
+        }
     }
 
     private void setUpInjections() {
@@ -110,12 +145,7 @@ public class ServerFileImageActivity extends AppCompatActivity implements
     private void setUpFullScreen() {
         final FullScreenHelper fullScreen = new FullScreenHelper(getSupportActionBar(), getImagePager());
         fullScreen.enableOnClickToggle(false);
-        getImagePager().setOnViewPagerClickListener(new ClickableViewPager.OnClickListener() {
-            @Override
-            public void onViewPagerClick(ViewPager viewPager) {
-                fullScreen.toggle();
-            }
-        });
+        getImagePager().setOnViewPagerClickListener(viewPager -> fullScreen.toggle());
         fullScreen.init();
     }
 
@@ -127,8 +157,8 @@ public class ServerFileImageActivity extends AppCompatActivity implements
     private void setUpImage() {
         setUpImageTitle();
         setUpImageAdapter();
-        setUpImagePosition();
         setUpImageListener();
+        setUpImagePosition();
     }
 
     private boolean isCastConnected() {
@@ -160,7 +190,7 @@ public class ServerFileImageActivity extends AppCompatActivity implements
     }
 
     private ClickableViewPager getImagePager() {
-        return (ClickableViewPager) findViewById(R.id.pager_images);
+        return findViewById(R.id.pager_images);
     }
 
     private ServerShare getShare() {
@@ -207,6 +237,43 @@ public class ServerFileImageActivity extends AppCompatActivity implements
         if (isCastConnected()) {
             loadRemoteMedia();
         }
+
+        setUpRecentFiles(getFiles().get(position));
+    }
+
+    private void setUpRecentFiles(ServerFile serverFile) {
+        long size;
+        if (isFileAvailableOffline(serverFile)) {
+            size = new File(getOfflineFileUri(serverFile.getName())).length();
+        } else {
+            size = serverFile.getSize();
+        }
+
+        String serverName = Preferences.getServerName(this);
+
+        RecentFile recentFile = new RecentFile(serverFile.getUniqueKey(),
+            getImageUri(serverFile),
+            serverName,
+            System.currentTimeMillis(),
+            size);
+        RecentFileRepository recentFileRepository = new RecentFileRepository(this);
+        recentFileRepository.insert(recentFile);
+    }
+
+    private boolean isFileAvailableOffline(ServerFile serverFile) {
+        OfflineFileRepository repository = new OfflineFileRepository(this);
+        OfflineFile file = repository.getOfflineFile(serverFile.getName(), serverFile.getModificationTime().getTime());
+        return file != null && file.getState() == OfflineFile.DOWNLOADED;
+    }
+
+    private String getUriFrom(String name, Date modificationTime) {
+        OfflineFileRepository repository = new OfflineFileRepository(this);
+        OfflineFile offlineFile = repository.getOfflineFile(name, modificationTime.getTime());
+        return offlineFile.getFileUri();
+    }
+
+    private String getOfflineFileUri(String name) {
+        return (getFilesDir() + "/" + Downloader.OFFLINE_PATH + "/" + name);
     }
 
     @Override
@@ -225,7 +292,7 @@ public class ServerFileImageActivity extends AppCompatActivity implements
                 return true;
 
             case R.id.menu_share:
-                startFileSharingActivity();
+                prepareDownload();
                 return true;
 
             default:
@@ -237,6 +304,28 @@ public class ServerFileImageActivity extends AppCompatActivity implements
         startFileDownloading(getShare(), getCurrentFile());
     }
 
+    private void prepareDownload() {
+
+        ServerFile serverFile = getCurrentFile();
+
+        if (isFileAvailableOffline(serverFile)) {
+            prepareDownloadingFile(serverFile);
+        } else {
+            startFileDownloading(getShare(), getCurrentFile());
+        }
+    }
+
+    private void prepareDownloadingFile(ServerFile file) {
+        PrepareDialogFragment fragment = new PrepareDialogFragment();
+        fragment.show(getSupportFragmentManager(), "prepare_dialog");
+        FileManager fm = FileManager.newInstance(this);
+        Uri offlinePath = fm.getContentUriForOfflineFile(file.getName());
+        File sourceLocation = new File(offlinePath.toString());
+        File downloadLocation = new File(getExternalFilesDir(Environment.DIRECTORY_DOWNLOADS), file.getName());
+
+        fm.copyFile(sourceLocation, downloadLocation);
+    }
+
     private ServerFile getCurrentFile() {
         return getImageFiles().get(getImagePager().getCurrentItem());
     }
@@ -246,7 +335,7 @@ public class ServerFileImageActivity extends AppCompatActivity implements
     }
 
     private void showFileDownloadingFragment(ServerShare share, ServerFile file) {
-        DialogFragment fragment = ServerFileDownloadingFragment.newInstance(share, file);
+        DialogFragment fragment = ServerFileDownloadingFragment.newInstance(share, file, FileOption.SHARE);
         fragment.show(getFragmentManager(), ServerFileDownloadingFragment.TAG);
     }
 
@@ -262,6 +351,21 @@ public class ServerFileImageActivity extends AppCompatActivity implements
     private void startFileSharingActivity(ServerFile file, Uri fileUri) {
         Intent intent = Intents.Builder.with(this).buildServerFileSharingIntent(file, fileUri);
         startActivity(intent);
+    }
+
+    @Subscribe
+    public void onFileCopied(FileCopiedEvent event) {
+        Uri contentUri = FileManager.newInstance(this).getContentUri(event.getTargetLocation());
+
+        dismissPreparingDialog();
+        finishFileDownloading(contentUri);
+    }
+
+    private void dismissPreparingDialog() {
+        PrepareDialogFragment fragment = (PrepareDialogFragment) getSupportFragmentManager().findFragmentByTag("prepare_dialog");
+        if (fragment != null && fragment.isAdded()) {
+            fragment.dismiss();
+        }
     }
 
     @Override
@@ -335,7 +439,9 @@ public class ServerFileImageActivity extends AppCompatActivity implements
     private void loadRemoteMedia() {
         final RemoteMediaClient remoteMediaClient = mCastSession.getRemoteMediaClient();
         if (remoteMediaClient != null) {
-            remoteMediaClient.load(buildMediaInfo());
+
+            MediaLoadOptions mediaLoadOptions = new MediaLoadOptions.Builder().build();
+            remoteMediaClient.load(buildMediaInfo(), mediaLoadOptions);
         }
     }
 
@@ -343,11 +449,36 @@ public class ServerFileImageActivity extends AppCompatActivity implements
         ServerFile file = getImageFiles().get(imagePosition);
         MediaMetadata imageMetadata = new MediaMetadata(MediaMetadata.MEDIA_TYPE_PHOTO);
         imageMetadata.putString(MediaMetadata.KEY_TITLE, file.getNameOnly());
-        String imageUrl = serverClient.getFileUri(getShare(), file).toString();
+        imageMetadata.putString(MediaMetadata.KEY_ARTIST, "");
+        imageMetadata.putString(MediaMetadata.KEY_ALBUM_TITLE, "");
+        String imageUrl = getImageUri(getCurrentFile());
+        imageMetadata.addImage(new WebImage(Uri.parse(imageUrl)));
+        imageMetadata.addImage(new WebImage(Uri.parse(imageUrl)));
         return new MediaInfo.Builder(imageUrl)
             .setStreamType(MediaInfo.STREAM_TYPE_BUFFERED)
             .setContentType(file.getMime())
             .setMetadata(imageMetadata)
             .build();
+    }
+
+    private String getImageUri(ServerFile serverFile) {
+        int fileType = getIntent().getIntExtra(Intents.Extras.FILE_TYPE, FileManager.SERVER_FILE);
+
+        if (fileType == FileManager.RECENT_FILE) {
+            return getRecentFileUri();
+        }
+        if (getShare() == null) {
+            OfflineFileRepository repository = new OfflineFileRepository(this);
+            OfflineFile offlineFile = repository.getOfflineFile(serverFile.getName(), serverFile.getModificationTime().getTime());
+            if (offlineFile != null) {
+                return offlineFile.getFileUri();
+            }
+        }
+        return serverClient.getFileUri(getShare(), getCurrentFile()).toString();
+    }
+
+    private String getRecentFileUri() {
+        RecentFileRepository repository = new RecentFileRepository(this);
+        return repository.getRecentFile(getCurrentFile().getUniqueKey()).getUri();
     }
 }
