@@ -22,6 +22,7 @@ package org.amahi.anywhere.fragment;
 import android.os.Bundle;
 import android.os.Parcelable;
 import android.support.v4.app.Fragment;
+import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.widget.DividerItemDecoration;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -29,6 +30,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.LinearLayout;
+import android.widget.Toast;
 
 import com.squareup.otto.Subscribe;
 
@@ -37,6 +39,8 @@ import org.amahi.anywhere.R;
 import org.amahi.anywhere.adapter.ServerSharesAdapter;
 import org.amahi.anywhere.bus.BusProvider;
 import org.amahi.anywhere.bus.ServerConnectionChangedEvent;
+import org.amahi.anywhere.bus.ServerConnectionFailedEvent;
+import org.amahi.anywhere.bus.ServerReconnectEvent;
 import org.amahi.anywhere.bus.ServerSharesLoadFailedEvent;
 import org.amahi.anywhere.bus.ServerSharesLoadedEvent;
 import org.amahi.anywhere.server.client.ServerClient;
@@ -51,7 +55,9 @@ import javax.inject.Inject;
 /**
  * Shares fragment. Shows shares list.
  */
-public class ServerSharesFragment extends Fragment {
+public class ServerSharesFragment extends Fragment implements
+    SwipeRefreshLayout.OnRefreshListener {
+
     @Inject
     ServerClient serverClient;
     private RecyclerView mRecyclerView;
@@ -62,26 +68,37 @@ public class ServerSharesFragment extends Fragment {
 
     private LinearLayout mErrorLinearLayout;
 
+    private SwipeRefreshLayout mSwipeRefreshLayout;
+
     @Override
     public View onCreateView(LayoutInflater layoutInflater, ViewGroup container, Bundle savedInstanceState) {
 
         View rootView = layoutInflater.inflate(R.layout.fragment_server_shares, container, false);
 
+        mSwipeRefreshLayout = (SwipeRefreshLayout) rootView.findViewById(R.id.layout_refresh);
+
         mRecyclerView = (RecyclerView) rootView.findViewById(R.id.list);
 
         mServerSharesAdapter = new ServerSharesAdapter(getActivity());
 
-        mRecyclerView.setLayoutManager(new LinearLayoutManager(getActivity(), LinearLayoutManager.VERTICAL, false));
+        mRecyclerView.setLayoutManager(new LinearLayoutManager(getActivity()
+            , LinearLayoutManager.VERTICAL, false));
 
         mEmptyLinearLayout = (LinearLayout) rootView.findViewById(R.id.empty);
 
         mErrorLinearLayout = (LinearLayout) rootView.findViewById(R.id.error);
+
+        setSwipeToRefresh();
 
         mRecyclerView.addItemDecoration(new
             DividerItemDecoration(getActivity(),
             DividerItemDecoration.VERTICAL));
 
         return rootView;
+    }
+
+    private void setSwipeToRefresh() {
+        mSwipeRefreshLayout.setOnRefreshListener(this);
     }
 
     @Override
@@ -153,12 +170,35 @@ public class ServerSharesFragment extends Fragment {
     }
 
     @Subscribe
+    public void onServerConnectionFailed(ServerConnectionFailedEvent event) {
+        Toast.makeText(getContext(), getResources()
+            .getString(R.string.message_error_amahi_anywhere_app), Toast.LENGTH_LONG).show();
+        showConnectionError();
+    }
+
+    private void showConnectionError() {
+        ViewDirector.of(getActivity(), R.id.animator).show(R.id.error);
+        mErrorLinearLayout.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                ViewDirector.of(getActivity(), R.id.animator).show(android.R.id.progress);
+                retryServerConnection();
+            }
+        });
+    }
+
+    private void retryServerConnection() {
+        BusProvider.getBus().post(new ServerReconnectEvent());
+    }
+
+    @Subscribe
     public void onServerConnectionChanged(ServerConnectionChangedEvent event) {
-        serverClient.getShares();
+        setUpSharesContent();
     }
 
     @Subscribe
     public void onSharesLoaded(ServerSharesLoadedEvent event) {
+        mSwipeRefreshLayout.setRefreshing(false);
         setUpSharesContent(event.getServerShares());
         showSharesContent();
     }
@@ -169,6 +209,7 @@ public class ServerSharesFragment extends Fragment {
     }
 
     private void showSharesError() {
+        mSwipeRefreshLayout.setRefreshing(false);
         ViewDirector.of(getActivity(), R.id.animator).show(R.id.error);
         mErrorLinearLayout.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -208,6 +249,11 @@ public class ServerSharesFragment extends Fragment {
 
     private boolean areSharesLoaded() {
         return getSharesAdapter() != null;
+    }
+
+    @Override
+    public void onRefresh() {
+        setUpSharesContent();
     }
 
     private static final class State {
