@@ -24,6 +24,7 @@ import android.content.Context;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
+import org.amahi.anywhere.BuildConfig;
 import org.amahi.anywhere.util.Time;
 
 import javax.inject.Singleton;
@@ -39,17 +40,29 @@ import retrofit2.converter.gson.GsonConverterFactory;
  * API dependency injection module. Provides resources such as HTTP client and JSON converter
  * for possible consumers.
  */
+
 @Module(
     complete = false,
     library = true
 )
 public class ApiModule {
+
+    /**
+     * After enabling the chuck dependencies modify the function definition below to pass ChuckInterceptor:
+     * provideHttpClient(ApiHeaders headers, HttpLoggingInterceptor logging, ChuckInterceptor chuck)
+     * <p>
+     * Add the Chuck interceptor when building OkHttpClient, using:
+     * clientBuilder.addInterceptor(chuck);
+     */
+
     @Provides
     @Singleton
     OkHttpClient provideHttpClient(ApiHeaders headers, HttpLoggingInterceptor logging) {
         OkHttpClient.Builder clientBuilder = new OkHttpClient.Builder();
+
         clientBuilder.addInterceptor(headers);
         clientBuilder.addInterceptor(logging);
+
         return clientBuilder.build();
     }
 
@@ -59,6 +72,17 @@ public class ApiModule {
         return new ApiHeaders(context);
     }
 
+    /**
+     * Creating an instance for ChuckInterceptor and providing it with context
+     * Uncomment the code below if using Chuck Interceptor for logging
+     */
+/*
+    @Provides
+    @Singleton
+    ChuckInterceptor provideChuckInterceptor(Context context) {
+        return new ChuckInterceptor(context);
+    }
+*/
     @Provides
     @Singleton
     Converter.Factory provideJsonConverterFactory(Gson json) {
@@ -74,7 +98,15 @@ public class ApiModule {
     @Provides
     @Singleton
     HttpLoggingInterceptor provideLogging() {
-        // change the level below to HttpLoggingInterceptor.Level.BODY to get the whole body in the logs
-        return new HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.HEADERS);
+        HttpLoggingInterceptor loggingInterceptor = new HttpLoggingInterceptor();
+
+        if (BuildConfig.DEBUG) {
+            // This level can be decreased to Level.HEADERS or Level.BASIC to reduce the details shown
+            loggingInterceptor.setLevel(HttpLoggingInterceptor.Level.BODY);
+        } else {
+            loggingInterceptor.setLevel(HttpLoggingInterceptor.Level.NONE);
+        }
+
+        return loggingInterceptor;
     }
 }
