@@ -20,7 +20,7 @@
 package org.amahi.anywhere.server.response;
 
 import org.amahi.anywhere.bus.BusProvider;
-import org.amahi.anywhere.bus.ServerConnectionFailedEvent;
+import org.amahi.anywhere.bus.ServerRouteLoadFailedEvent;
 import org.amahi.anywhere.bus.ServerRouteLoadedEvent;
 import org.amahi.anywhere.server.model.ServerRoute;
 
@@ -44,6 +44,6 @@ public class ServerRouteResponse implements Callback<ServerRoute> {
 
     @Override
     public void onFailure(Call<ServerRoute> call, Throwable t) {
-        BusProvider.getBus().post(new ServerConnectionFailedEvent());
+        BusProvider.getBus().post(new ServerRouteLoadFailedEvent(t.getMessage()));
     }
 }
