@@ -24,6 +24,7 @@ import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
 
 import org.amahi.anywhere.R;
+import org.amahi.anywhere.fragment.ServerFilesFragment;
 
 import static android.content.Context.MODE_PRIVATE;
 
@@ -44,7 +45,6 @@ public final class Preferences {
     public static String getServerConnection(SharedPreferences preferences, Context context) {
         return preferences.getString(context.getString(R.string.preference_key_server_connection), context.getString(R.string.preference_entry_server_connection_auto));
     }
-
 
     public static void setPrefAuto(SharedPreferences preference, Context context) {
         preference.edit().putString(context.getString(R.string.preference_key_server_connection), context.getString(R.string.preference_entry_server_connection_auto)).apply();
@@ -86,6 +86,14 @@ public final class Preferences {
         context.getSharedPreferences(context.getString(R.string.preference), MODE_PRIVATE).edit().putBoolean(context.getString(R.string.is_first_run), false).apply();
     }
 
+    public static void setServerName(Context context, String serverName) {
+        context.getSharedPreferences(context.getString(R.string.preference), MODE_PRIVATE).edit().putString(context.getString(R.string.pref_server_select_key), serverName).apply();
+    }
+
+    public static String getServerName(Context context) {
+        return context.getSharedPreferences(context.getString(R.string.preference), MODE_PRIVATE).getString(context.getString(R.string.pref_server_select_key), null);
+    }
+
     public static Preferences ofCookie(Context context) {
         return new Preferences(context, Locations.COOKIE);
     }
@@ -106,6 +114,14 @@ public final class Preferences {
         preferences.edit().putString(key, value).apply();
     }
 
+    public static int getSortOption(Context context) {
+        return getPreference(context).getInt(Defaults.SORTING_OPTION, ServerFilesFragment.SORT_MODIFICATION_TIME);
+    }
+
+    public static void setSortOption(Context context, int filesSort) {
+        getPreference(context).edit().putInt(Defaults.SORTING_OPTION, filesSort).apply();
+    }
+
     private static final class Locations {
         public static final String COOKIE = "cookie";
 
@@ -115,6 +131,7 @@ public final class Preferences {
 
     private static final class Defaults {
         public static final String STRING = "";
+        public static final String SORTING_OPTION = "sorting_option";
 
         private Defaults() {
         }
