@@ -109,7 +109,7 @@ public class UploadManager {
 
     private String getUploadPath() {
         return PreferenceManager.getDefaultSharedPreferences(context)
-            .getString(context.getString(R.string.preference_key_upload_path), null);
+            .getString(context.getString(R.string.preference_key_upload_location), null);
     }
 
 
@@ -132,12 +132,7 @@ public class UploadManager {
         }
 
         final Handler handler = new Handler();
-        handler.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                processNextFile();
-            }
-        }, 500);
+        handler.postDelayed(this::processNextFile, 500);
     }
 
     public interface UploadCallbacks {
